# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "3.0.0",
    "title": "X2Mint API",
    "description": "x2mint-api"
  },
  "servers": [
    {
      "url": "http://localhost:5005/"
    }
  ],
  "paths": {
...
    "/app/api/v1/tests/{testId}": {
      "get": {
        "tags": [
          "tests"
        ],
        "summary": "Get test by id",
        "description": "",
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "tests"
        ],
        "summary": "Update a test by test Id by admin/creator",
        "description": "",
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "verifyAccount": {
                    "example": "any"
                  },
                  "name": {
                    "example": "any"
                  },
                  "creatorId": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "questions": {
                    "example": "any"
                  },
                  "pin": {
                    "example": "any"
                  },
                  "startTime": {
                    "example": "any"
                  },
                  "endTime": {
                    "example": "any"
                  },
                  "maxPoints": {
                    "example": "any"
                  },
                  "maxTimes": {
                    "example": "any"
                  },
                  "_status": {
                    "example": "any"
                  },
                  "questionsOrder": {
                    "example": "any"
                  },
                  "tracking": {
                    "example": "any"
                  },
                  "userAuth": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/app/api/v1/contests/{contestId}/tests": {
      "get": {
        "tags": [
          "contests"
        ],
        "summary": "Get all tests by contest id for admin/creator",
        "description": "",
        "parameters": [
          {
            "name": "contestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "contests"
        ],
        "summary": "Update/create new tests for contest by admin/creator",
        "description": "",
        "parameters": [
          {
            "name": "contestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "verifyAccount": {
                    "example": "any"
                  },
                  "tests": {
                    "example": "any"
                  },
                  "userAuth": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/api/v1/contests/{contestId}/taketests": {
      "get": {
        "tags": [
          "contests"
        ],
        "summary": "Get all takeTests by contest id for admin/creator",
        "description": "",
        "parameters": [
          {
            "name": "contestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/app/api/v1/takeTest/{takeTestId}": {
      "get": {
        "tags": [
          "takeTest"
        ],
        "summary": "Get take test by id",
        "description": "",
        "parameters": [
          {
            "name": "takeTestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "takeTest"
        ],
        "summary": "Update take test by admin/creator/user",
        "description": "",
        "parameters": [
          {
            "name": "takeTestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "verifyAccount": {
                    "example": "any"
                  },
                  "test": {
                    "example": "any"
                  },
                  "questionsOrder": {
                    "example": "any"
                  },
                  "chooseAnswers": {
                    "example": "any"
                  },
                  "_status": {
                    "example": "any"
                  },
                  "action": {
                    "example": "any"
                  },
                  "userAuth": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/app/api/v1/takeTest/test/{testId}/{userId}": {
      "get": {
        "tags": [
          "takeTest"
        ],
        "summary": "Get all take test by testId and userId",
        "description": "",
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/app/api/v1/takeTest/test/{testId}/{userId}/lastest": {
      "get": {
        "tags": [
          "takeTest"
        ],
        "summary": "Get lastest take test by testId and userId",
        "description": "",
        "parameters": [
          {
            "name": "testId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/app/api/v1/takeTest/{takeTestId}/logs": {
      "get": {
        "tags": [
          "takeTest"
        ],
        "summary": "Get take test logs by id",
        "description": "",
        "parameters": [
          {
            "name": "takeTestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/app/api/v1/takeTest/user/{userId}": {
      "get": {
        "tags": [
          "takeTest"
        ],
        "summary": "Get all take test by userId",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}